backend/pyproject.toml
#./backend/pyproject.toml
# Project metadata
[tool.poetry]
name = "azlon"
version = "0.0.1"
description = "autonomous coding project solver"
authors = [
    "Harrison E. Muchnic <hem9984@nyu.edu>",
]
readme = "readme.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
restack-ai = "0.0.50"
pydantic = "^2.10.3"
fastapi = "0.115.4"  
uvicorn = "^0.22.0"
baml-py = "^0.77.0"
kaggle = "^1.6.17"
mem0ai = "^0.1.56"
e2b-code-interpreter = "^1.0.5"
#e2b-cli = "^0.1.9"
e2b = "^1.1.0"
boto3 = "^1.37.6"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
autoflake = "^2.2.1"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

# Build system configuration
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Code style tools configuration
[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true
remove-duplicate-keys = true
in-place = true
recursive = true
expand-star-imports = true
check = false

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

# CLI command configuration
[tool.poetry.scripts]
services = "src.services:run_services"
schedule = "schedule_workflow:run_schedule_workflow"